CLE/requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
astroid==2.0.3
async-generator==1.9
attrs==18.1.0
Automat==0.7.0
awscli==1.15.59
beautifulsoup4==4.6.3
billiard==3.5.0.4
boto3==1.7.65
botocore==1.10.58
bs4==0.0.1
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
chartjs==1.2
chromedriver-binary==2.42.0
colorama==0.3.9
constantly==15.1.0
coverage==3.6
cryptography==2.2.1
cssselect==1.0.3
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-background-tasks==1.2.0
django-celery-beat==1.1.1
django-compat==1.0.15
django-formtools==2.1
django-widget-tweaks==1.4.3
dnspython==1.15.0
docutils==0.14
eventlet==0.24.1
ez-setup==0.9
future==0.16.0
greenlet==0.4.14
hyperlink==17.3.1
idna==2.7
incremental==17.5.0
ipgetter==0.7
isort==4.3.4
jmespath==0.9.3
kombu==4.2.1
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.1.0
mysql-connector-python==8.0.11
mysqlclient==1.3.10
oauthlib==2.1.0
parsel==1.4.0
phonenumbers==8.9.11
pluggy==0.6.0
protobuf==3.6.0
py==1.5.3
pyaes==1.6.1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodome==3.6.6
PyDispatcher==2.0.5
pylint==2.1.1
pyOpenSSL==18.0.0
pytest==3.5.1
pytest-runner==4.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytyp==2.2.4
pytz==2018.5
pywin32==221
PyYAML==3.13
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
Scrapy==1.5.0
selenium==3.14.1
service-identity==17.0.0
six==1.11.0
sqlparse==0.1.16
telegram==0.0.1
Telethon==1.1.1
Telethon-sync==1.0.4
Twisted==17.5.0
typed-ast==1.1.0
urllib3==1.23
vine==1.1.4
w3lib==1.19.0
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
zope.interface==4.5.0
